Move service commands into its own namespace
